requirements.txt
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
cachelib==0.1.1
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
click==7.1.2
colorama==0.4.4
cryptography==38.0.1
distlib==0.3.0
Django==3.1.4
djangorestframework==3.14.0
filelock==3.0.12
Flask==1.1.2
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.3
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
markdown2==2.3.10
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycparser==2.21
pylint==2.6.0
pytz==2020.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.6
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.10
virtualenv==20.0.21
Werkzeug==1.0.1
wrapt==1.12.1
